+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -34,9 +34,4 @@ public class StationController {
         stationService.deleteStationById(id);
         return ResponseEntity.noContent().build();
     }
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity handleIllegalArgsException(final DataIntegrityViolationException e) {
-        return ResponseEntity.badRequest().build();
-    }
 }
